requirements.txt
streamlit>=1.30.0
openai>=1.10.0
anthropic>=0.21.0
supabase>=2.3.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
